+++ grid2op/Environment/_ObsEnv.py
@@ -96,25 +96,12 @@ class _ObsEnv(BaseEnv):
         # initialize the observation space
         self._obsClass = None
 
-        # self.gen_activeprod_t_init = np.zeros(self.n_gen, dtype=dt_float)
-        # self.gen_activeprod_t_redisp_init = np.zeros(self.n_gen, dtype=dt_float)
-        # self.times_before_line_status_actionable_init = np.zeros(
-        #     self.n_line, dtype=dt_int
-        # )
-        # self.times_before_topology_actionable_init = np.zeros(self.n_sub, dtype=dt_int)
-        # self.time_next_maintenance_init = np.zeros(self.n_line, dtype=dt_int)
-        # self.duration_next_maintenance_init = np.zeros(self.n_line, dtype=dt_int)
-        # self.target_dispatch_init = np.zeros(self.n_gen, dtype=dt_float)
-        # self.actual_dispatch_init = np.zeros(self.n_gen, dtype=dt_float)
-        # self._already_modified_gen_init = np.zeros(self.n_gen, dtype=dt_float)
-
         # line status (inherited from BaseEnv)
         self._line_status = np.full(self.n_line, dtype=dt_bool, fill_value=True)
         # line status (for this usage)
         self._line_status_me = np.ones(
             shape=self.n_line, dtype=dt_int
         )  # this is "line status" but encode in +1 / -1
-        # self._line_status_orig = np.ones(shape=self.n_line, dtype=dt_int)
 
         if self._thermal_limit_a is None:
             self._thermal_limit_a = 1.0 * thermal_limit_a.astype(dt_float)
@@ -145,9 +132,6 @@ class _ObsEnv(BaseEnv):
         ####
 
         self.no_overflow_disconnection = parameters.NO_OVERFLOW_DISCONNECTION
-
-        # self._load_p, self._load_q, self._load_v = None, None, None
-        # self._prod_p, self._prod_q, self._prod_v = None, None, None
         self._topo_vect = np.zeros(type(backend_instanciated).dim_topo, dtype=dt_int)
 
         # other stuff
@@ -160,50 +144,6 @@ class _ObsEnv(BaseEnv):
             self._backend_action_set = None
         else:
             self._backend_action_set = self._backend_action_class()
-        
-        # opponent
-        # self.opp_space_state = None
-        # self.opp_state = None
-
-        # storage
-        # self._storage_current_charge_init = None
-        # self._storage_previous_charge_init = None
-        # self._action_storage_init = None
-        # self._amount_storage_init = None
-        # self._amount_storage_prev_init = None
-        # self._storage_power_init = None
-
-        # # storage unit
-        # if self.__unusable:
-        #     self._storage_current_charge_init = np.zeros(0, dtype=dt_float)
-        #     self._storage_previous_charge_init = np.zeros(0, dtype=dt_float)
-        #     self._action_storage_init = np.zeros(0, dtype=dt_float)
-        #     self._storage_power_init = np.zeros(0, dtype=dt_float)
-        # else:
-        #     self._storage_current_charge_init = np.zeros(self.n_storage, dtype=dt_float)
-        #     self._storage_previous_charge_init = np.zeros(self.n_storage, dtype=dt_float)
-        #     self._action_storage_init = np.zeros(self.n_storage, dtype=dt_float)
-        #     self._storage_power_init = np.zeros(self.n_storage, dtype=dt_float)
-            
-        # self._amount_storage_init = 0.0
-        # self._amount_storage_prev_init = 0.0
-            
-
-        # # curtailment
-        # if self.__unusable:
-        #     self._limit_curtailment_init = np.zeros(0, dtype=dt_float)
-        #     self._gen_before_curtailment_init = np.zeros(0, dtype=dt_float)
-        # else:
-        #     self._limit_curtailment_init = np.zeros(self.n_gen, dtype=dt_float)
-        #     self._gen_before_curtailment_init = np.zeros(self.n_gen, dtype=dt_float)
-        # self._sum_curtailment_mw_init = 0.0
-        # self._sum_curtailment_mw_prev_init = 0.0
-
-        # # step count
-        # self._nb_time_step_init = 0
-
-        # # alarm / attention budget
-        # self._attention_budget_state_init = None
 
         if self.__unusable:
             self._disc_lines = np.zeros(shape=0, dtype=dt_int) - 1
@@ -366,26 +306,32 @@ class _ObsEnv(BaseEnv):
         
         self._reset_to_orig_state(obs)
         self._topo_vect[:] = obs.topo_vect
-        (
-            still_in_maintenance,
-            reconnected,
-            first_ts_maintenance,
-        ) = self._update_vector_with_timestep(time_step)
-        if np.any(first_ts_maintenance):
-            set_status = np.array(self._line_status_me, dtype=dt_int)
-            set_status[first_ts_maintenance] = -1
-            topo_vect = np.array(self._topo_vect, dtype=dt_int)
-            topo_vect[self.line_or_pos_topo_vect[first_ts_maintenance]] = -1
-            topo_vect[self.line_ex_pos_topo_vect[first_ts_maintenance]] = -1
+        
+        if time_step >= 1:
+            # handle the compoenents that depends on the time
+            (
+                still_in_maintenance,
+                reconnected,
+                first_ts_maintenance,
+            ) = self._update_vector_with_timestep(time_step)
+            if np.any(first_ts_maintenance):
+                set_status = np.array(self._line_status_me, dtype=dt_int)
+                set_status[first_ts_maintenance] = -1
+                topo_vect = np.array(self._topo_vect, dtype=dt_int)
+                topo_vect[self.line_or_pos_topo_vect[first_ts_maintenance]] = -1
+                topo_vect[self.line_ex_pos_topo_vect[first_ts_maintenance]] = -1
+            else:
+                set_status = self._line_status_me
+                topo_vect = self._topo_vect
+            
+            if np.any(still_in_maintenance):
+                set_status[still_in_maintenance] = -1
+                topo_vect = np.array(self._topo_vect, dtype=dt_int)
+                topo_vect[self.line_or_pos_topo_vect[still_in_maintenance]] = -1
+                topo_vect[self.line_ex_pos_topo_vect[still_in_maintenance]] = -1
         else:
             set_status = self._line_status_me
             topo_vect = self._topo_vect
-        
-        if np.any(still_in_maintenance):
-            set_status[still_in_maintenance] = -1
-            topo_vect = np.array(self._topo_vect, dtype=dt_int)
-            topo_vect[self.line_or_pos_topo_vect[still_in_maintenance]] = -1
-            topo_vect[self.line_ex_pos_topo_vect[still_in_maintenance]] = -1
             
         # TODO set the shunts here
         # update the action that set the grid to the real value
@@ -469,10 +415,6 @@ class _ObsEnv(BaseEnv):
         first_ts_maintenance = tnm_orig == time_step
         still_in_maintenance = maint_started & (~maint_over) & (~first_ts_maintenance)
         
-        if hasattr(self, "_DEBUG") and self._DEBUG: 
-            import pdb
-            pdb.set_trace()
-            
         # count down time next maintenance
         self._time_next_maintenance[:] = np.maximum(
             self._time_next_maintenance - time_step, -1
@@ -484,7 +426,6 @@ class _ObsEnv(BaseEnv):
 
         # powerline that will be in maintenance at this time step
         self._time_next_maintenance[first_ts_maintenance] = 0
-        # self._duration_next_maintenance[first_ts_maintenance] -= time_step
         
         # powerline that will be in maintenance at this time step
         self._time_next_maintenance[reconnected | maint_over] = -1
@@ -509,8 +450,9 @@ class _ObsEnv(BaseEnv):
         """
         self.reset()  # reset the "BaseEnv"
         self.backend.set_thermal_limit(obs._thermal_limit)
-        self._oppSpace._set_state(obs._env_internal_params["opp_space_state"], 
-                                  obs._env_internal_params["opp_state"])
+        if "opp_space_state" in obs._env_internal_params:
+            self._oppSpace._set_state(obs._env_internal_params["opp_space_state"], 
+                                      obs._env_internal_params["opp_state"])
         # storage unit
         self._storage_current_charge[:] = obs.storage_charge
         self._storage_previous_charge[:] = obs._env_internal_params["_storage_previous_charge"]
@@ -606,7 +548,6 @@ class _ObsEnv(BaseEnv):
         maybe_exc = self._ptr_orig_obs_space.can_use_simulate()
         if maybe_exc is not None:
             raise maybe_exc
-        
         obs, reward, done, info = self.step(action)
         return obs, reward, done, info
 
@@ -652,67 +593,6 @@ class _ObsEnv(BaseEnv):
                            "environment that cannot be copied.")
         self.is_init = False
         
-        # real_backend = env.backend
-
-        # self._load_p, self._load_q, self._load_v = real_backend.loads_info()
-        # self._prod_p, self._prod_q, self._prod_v = real_backend.generators_info()
-        # self._topo_vect = real_backend.get_topo_vect()
-
-        # convert line status to -1 / 1 instead of false / true
-        # self._line_status_orig[:] = env.get_current_line_status().astype(
-        #     dt_int
-        # )  # false -> 0 true -> 1
-        # self._line_status_orig *= 2  # false -> 0 true -> 2
-        # self._line_status_orig -= 1  # false -> -1; true -> 1
-        # self.is_init = False
-
-        # Make a copy of env state for simulation
-        # TODO this depends on the datetime simulated, so find a way to have it independant of that !!!
-        # if self._thermal_limit_a is None:
-        #     self._thermal_limit_a = 1.0 * env._thermal_limit_a.astype(dt_float)
-        # else:
-        #     self._thermal_limit_a[:] = env._thermal_limit_a.astype(dt_float)
-        # self.gen_activeprod_t_init[:] = env._gen_activeprod_t
-        # self.gen_activeprod_t_redisp_init[:] = env._gen_activeprod_t_redisp
-        # self.times_before_line_status_actionable_init[
-        #     :
-        # ] = env._times_before_line_status_actionable
-        # self.times_before_topology_actionable_init[
-        #     :
-        # ] = env._times_before_topology_actionable
-        # self.time_next_maintenance_init[:] = env._time_next_maintenance
-        # self.duration_next_maintenance_init[:] =  env._duration_next_maintenance
-        # self.target_dispatch_init[:] = env._target_dispatch
-        # self.actual_dispatch_init[:] = env._actual_dispatch
-        # self._already_modified_gen_init[:] = env._already_modified_gen
-        # self.opp_space_state, self.opp_state = env._oppSpace._get_state()
-
-        # storage units
-        # TODO this is not time independant... i set up the previous charge of the obs env to be
-        # set current charge of the simulated env on purpose
-        # self._storage_current_charge_init[:] = env._storage_current_charge
-        # self._storage_previous_charge_init[:] = env._storage_previous_charge
-        # self._action_storage_init[:] = env._action_storage
-        # self._amount_storage_init = 1.0 * env._amount_storage
-        # self._amount_storage_prev_init = 1.0 * env._amount_storage_prev
-        # self._storage_power_init[:] = env._storage_power
-
-        # curtailment
-        # self._limit_curtailment_init[:] = env._limit_curtailment
-        # self._gen_before_curtailment_init[:] = env._gen_before_curtailment
-        # self._sum_curtailment_mw_init = 1.0 * env._sum_curtailment_mw
-        # self._sum_curtailment_mw_prev_init = 1.0 * env._sum_curtailment_mw_prev
-
-        # # time delta
-        # self.delta_time_seconds = env.delta_time_seconds
-
-        # # current time
-        # self._nb_time_step_init = env.nb_time_step
-
-        # # attention budget
-        # if self._has_attention_budget:
-        #     self._attention_budget_state_init = env._attention_budget.get_state()
-
     def get_current_line_status(self):
         if self.__unusable:
             raise EnvError("Impossible to use a Observation backend with an "
@@ -731,39 +611,8 @@ class _ObsEnv(BaseEnv):
         # clean all the attributes
         for attr_nm in [
             "_obsClass",
-            # "gen_activeprod_t_init",
-            # "gen_activeprod_t_redisp_init",
-            # "times_before_line_status_actionable_init",
-            # "times_before_topology_actionable_init",
-            # "time_next_maintenance_init",
-            # "duration_next_maintenance_init",
-            # "target_dispatch_init",
             "_line_status",
             "_line_status_me",
-            # "_line_status_orig",
-            # "_load_p",
-            # "_load_q",
-            # "_load_v",
-            # "_prod_p",
-            # "_prod_q",
-            # "_prod_v",
-            # "_topo_vect",
-            # "opp_space_state",
-            # "opp_state",
-            # "_storage_current_charge_init",
-            # "_storage_previous_charge_init",
-            # "_action_storage_init",
-            # "_amount_storage_init",
-            # "_amount_storage_prev_init",
-            # "_storage_power_init",
-            # "_storage_current_charge_init",
-            # "_storage_previous_charge_init",
-            # "_limit_curtailment_init",
-            # "_gen_before_curtailment_init",
-            # "_sum_curtailment_mw_init",
-            # "_sum_curtailment_mw_prev_init",
-            # "_nb_time_step_init",
-            # "_attention_budget_state_init",
             "_max_episode_duration",
             "_ptr_orig_obs_space",
         ]: